requirements/dev.txt
-r common.txt
black==23.1.0
django-debug-toolbar==4.2.0
flake8==6.0.0
pep8-naming==0.13.3
flake8-quotes==3.3.2
flake8-import-order==0.18.2
